Engine+Game Android Port (#3)
* -android preperations

* - first working apk build (still much more in wip)

* - android iteration

* - updated object selection system to color picking (becouse stencil reading not supported on ES)

* - fix for selection on mobile, removed stecil test and updated androd project

* - improved cmake, removed garbage

* - removed another generated file

* - fix for android gitignores

* - game assets files now properly copies to android and no longer in SC

* - minor code cleanup

+++ dreco-engine/src/renderer/vertex_properties.hxx
@@ -13,7 +13,7 @@ struct vertex_properties
 
 	std::vector<float> vertexes = {};
 
-	std::vector<uint32_t> vert_elem = {};
+	std::vector<uint8_t> vert_elem = {};
 
 	std::vector<float> texture_coords = {};
